added: init_survey.java, activity_init_survey.xml, modified: MainActivity.java
implemented the basic UI and can now iterate throught the survey questions. Includes multiple choice buttons for each question, although these do not yet update to reflect the particular choices for each question. Also, modified MainActivity.java so that the app opens by default to the initialization survey. This is purely for testing purposes and will be undone later