modified: addactivity.java (monthyl bill feature)
upon adding a bill activity (electricity, gas, water), program automatically adds emissions related to the bill for the past 30 days (to simulate that the bill was a monthly bill for total energy use over the course of the past month). The emissions per day are the same for each of the past 30 days.            Also note; changes made to fetchSnapshot() method of ecotracker in last commit that I forgot to log in the commit message. Made this method static (along with its associated variables) in order to be able to call is from addactivity to update ecotracker automatically after an activity is added/edited