modified: addactivity.java and surveyfragment.java
added the feature where the survey question that asks what car the user drives saves the car as default_car in the user's firebase data. This default car is then retrieved in EcoTracker (as mentioned in requirements of the project) and displayed to the user if they try to add a drive personal vehicle activity.       Along with this, some slight restructuring of addactivity file. Nothing major.