modified: everything related to ecotracker and nav_graph
ecotracker now is the second button on the nav graph. Works fine, with one remaining issue to tackle; calendar day is not saved when we go to add an activity then return to eco tracker. It just resets to the current day. Need to fix this.          Also (low priority) consider removing all old, commented out code related to loadFragment() style of loading EcoTracker and AddActivity since this is now useless, we use navigate() via navigation bar.